[UPDATE] Makefile compiles now new get_name / get_color funcs
[UPDATE] Vm loop now updates map tiles

+++ corewar/src/graphic/update/map/map_tiles.c
@@ -39,8 +39,8 @@ void update_tile(sfVertexArray *tile, sfVector3f point, int color,
 
 void update_tile_map(map_formatter_t *map)
 {
-    register size_t x = 0;
-    register size_t y = 0;
+    register int x = 0;
+    register int y = 0;
 
     for (y = 0; y < map->map_settings.size.y; y += 1) {
         for (x = 0; x < map->map_settings.size.x; x += 1) {
